Add a special layout to generate the attributions index
The attributions index was provided pre-written in its own manuscript file. However, in keeping with Quire's principal that content should only ever exist in a single place in Markdown or YAML, we're instead autogenerating the index with this custom template from the data in the objects.yml file.

Once the objects.yml file is complete, the generated index will need to be double-checked for fidelity to the manuscript version.